can add group expenses, currently all group members must be part of it

+++ src/main/java/com/example/owetracker/model/Expense.java
@@ -24,7 +24,7 @@ public class Expense {
     private User paidBy;
 
     @Column(name = "group_id")
-    private Long groupId;
+    private Integer groupId;
 
     private String status;
 
@@ -44,7 +44,7 @@ public class Expense {
     public Expense() {
     }
 
-    public Expense(String title, String description, BigDecimal amount, User paidBy, Long groupId, String status, LocalDateTime createdAt, List<User> participants) {
+    public Expense(String title, String description, BigDecimal amount, User paidBy, Integer groupId, String status, LocalDateTime createdAt, List<User> participants) {
         this.title = title;
         this.description = description;
         this.amount = amount;
@@ -95,11 +95,11 @@ public class Expense {
         this.paidBy = paidBy;
     }
 
-    public Long getGroupId() {
+    public Integer getGroupId() {
         return groupId;
     }
 
-    public void setGroupId(Long groupId) {
+    public void setGroupId(Integer groupId) {
         this.groupId = groupId;
     }
